Extract concourse script formatting helpers to lib
So we can re-use
