+++ concourse/scripts/check-certificates.rb
@@ -4,27 +4,7 @@ require 'openssl'
 require 'json'
 require 'date'
 
-def separator
-  puts('-' * 80)
-end
-
-class String
-  def red
-    "\e[31m#{self}\e[0m"
-  end
-
-  def green
-    "\e[32m#{self}\e[0m"
-  end
-
-  def yellow
-    "\e[33m#{self}\e[0m"
-  end
-
-  def blue;
-    "\e[34m#{self}\e[0m"
-  end
-end
+require_relative './lib/formatting'
 
 class CredHubClient
   attr_reader :api_url
