+++ concourse/scripts/lib/formatting.rb
@@ -1,0 +1,21 @@
+def separator
+  puts('-' * 80)
+end
+
+class String
+  def red
+    "\e[31m#{self}\e[0m"
+  end
+
+  def green
+    "\e[32m#{self}\e[0m"
+  end
+
+  def yellow
+    "\e[33m#{self}\e[0m"
+  end
+
+  def blue;
+    "\e[34m#{self}\e[0m"
+  end
+end
